pyproject.toml
[tool.poetry]
name = "pubsub-wrap"
version = "0.1.0"
description = ""
authors = ["Dan <dntknwhtf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-pubsub = "^2.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
